+++ docs/changelog.md
@@ -10,6 +10,8 @@
 - [#343](https://github.com/estruyf/vscode-demo-time/issues/343): added the zoom
   actions: `zoomIn`, `zoomOut`, and `zoomReset` to control the editor zoom level
   during your demo
+- [`#383`](https://github.com/estruyf/vscode-demo-time/issues/383): Added the `enableZenMode` and `disableZenMode` actions to control Zen Mode
+  during Scenes
 - [#348](https://github.com/estruyf/vscode-demo-time/issues/348): Added the
   analytics PRO feature